backend/requirements.txt
flask==2.3.2
flask_cors==3.0.10
Pillow==9.5.0
